+++ .prettierrc.json
@@ -1,0 +1,20 @@
+{
+    "arrowParens": "always",
+    "bracketSameLine": false, 
+    "bracketSpacing": true, 
+    "embeddedLanguageFormatting": "auto", 
+    "htmlWhitespaceSensitivity": "css",
+    "insertPragma": false, 
+    "jsxSingleQuote": false, 
+    "printWidth": 80,
+    "proseWrap": "preserve", 
+    "quoteProps": "as-needed", 
+    "requirePragma": false, 
+    "semi": true, 
+    "singleAttributePerLine": false,
+    "singleQuote": false, 
+    "tabWidth": 4, 
+    "trailingComma": "es5",
+    "useTabs": false, 
+    "jsxBracketSameLine": false
+}
