+++ public/_redirects.html
@@ -1,0 +1,1 @@
+/* /index.html 200
