+++ src/components/Header.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Profile from "./api/profile";
+import Profile from "./profile";
 import "../styles/Header.css";
 
 function Header() {
